# perf: add some caching to symbolic