The functions in symbolic.py lend themselves to caching.
```
codegen         mean runtime:  214.38ms, runs:   271.73,  192.78,  187.83,  198.34,  214.70,  201.50,  207.98,  204.32,  199.40,  265.23
methodcache     mean runtime:  162.07ms, runs:   180.95,  179.14,  159.69,  150.42,  186.42,  152.14,  151.52,  152.13,  158.89,  149.41
profile         mean runtime:  922.64ms, runs:   801.41,  919.37,  865.20,  993.67,  910.14,  974.35,  969.17,  925.93,  922.24,  944.88

to
codegen         mean runtime:  194.27ms, runs:   267.85,  179.58,  189.86,  173.78,  174.06,  173.80,  182.35,  242.64,  187.00,  171.82
methodcache     mean runtime:  162.85ms, runs:   153.69,  166.74,  154.29,  170.61,  151.54,  150.07,  150.69,  153.29,  153.83,  223.79
profile         mean runtime:  722.39ms, runs:   695.66,  722.07,  701.50,  719.86,  747.23,  724.80,  723.51,  737.06,  730.08,  722.16


from
792.73 lazy.py
642.66 tensor.py
120.64 shapetracker.py
139.22 ops.py
356.79 symbolic.py
170.67 helpers.py
33.97 lib.py
2256.68 total

to
790.71 lazy.py
640.07 tensor.py
108.52 shapetracker.py
138.15 ops.py
104.15 symbolic.py
159.20 helpers.py
31.62 lib.py
1972.42 total


from
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    86667  298.065    0.003  925.243    0.011 symbolic.py:119(sum)
24005/13498  143.739    0.006 1038.990    0.077 symbolic.py:228(__floordiv__)
    38152  103.572    0.003  178.330    0.005 symbolic.py:296(create_rednode)
    85480   91.532    0.001  232.350    0.003 symbolic.py:62(__mul__)
59540/57463   82.742    0.001  288.487    0.005 symbolic.py:286(flat_components)
    83234   65.806    0.001  121.354    0.001 symbolic.py:178(__init__)
    10317   57.692    0.006  305.974    0.030 symbolic.py:260(__mod__)
   121386   52.057    0.000   62.982    0.001 symbolic.py:172(create_node)
    86667   39.694    0.000   39.694    0.000 symbolic.py:121(<listcomp>)
    33731   36.958    0.001   77.753    0.002 symbolic.py:131(<listcomp>)
    21721   32.006    0.001  106.025    0.005 symbolic.py:75(__floordiv__)
    17764   31.897    0.002  127.354    0.007 symbolic.py:47(__lt__)
    97365   30.935    0.000   30.935    0.000 symbolic.py:161(__init__)
   101527   29.263    0.000   74.657    0.001 symbolic.py:34(__hash__)
35992/13281   28.497    0.001  149.299    0.011 symbolic.py:18(render)
    44651   23.858    0.001   23.858    0.001 symbolic.py:201(get_bounds)
    26986   21.239    0.001   85.335    0.003 symbolic.py:194(__floordiv__)
    19526   20.768    0.001  153.510    0.008 symbolic.py:40(__add__)
    76089   20.484    0.000   20.484    0.000 symbolic.py:223(__init__)
     6102   19.415    0.003  195.030    0.032 symbolic.py:273(__lt__)

to
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
39583/18080   28.957    0.001  144.023    0.008 symbolic.py:18(render)
    83781   21.971    0.000   49.670    0.001 symbolic.py:34(__hash__)
    33764   20.799    0.001   66.556    0.002 symbolic.py:36(__eq__)
40320/20160   17.326    0.000   17.326    0.001 symbolic.py:173(__eq__)
     5772   14.585    0.003   67.160    0.012 symbolic.py:121(sum)
    19811   11.630    0.001   16.204    0.001 symbolic.py:316(sym_infer)
    29099   10.956    0.000   79.392    0.003 symbolic.py:174(__hash__)
    11029   10.747    0.001   63.460    0.006 symbolic.py:31(hash)
     8882    9.781    0.001   38.411    0.004 symbolic.py:46(__ge__)
     5760    9.750    0.002   35.667    0.006 symbolic.py:142(ands)
    17776    9.382    0.001   14.277    0.001 symbolic.py:146(<genexpr>)
     8882    7.712    0.001   26.475    0.003 symbolic.py:39(__neg__)
     6976    7.602    0.001   20.481    0.003 symbolic.py:321(<lambda>)
    11029    7.023    0.001   20.859    0.002 symbolic.py:29(key)
     7843    5.490    0.001    8.110    0.001 symbolic.py:110(num)
```