@tinyb0t 2023-09-04
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/shape/symbolic.py                15               5
------------------------------------------------------------
lines added in the tinygrad folder: 10
```