@geohot 2023-09-04
This isn't worth it, it's 11 lines.

If it's not improving readability, it won't be merged.